refactor: 프리뷰 함수 이동
