+++ app/src/main/java/com/sopt/now/compose/feature/main/MainActivity.kt
@@ -3,8 +3,6 @@ package com.sopt.now.compose.feature.main
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.tooling.preview.Preview
 import com.sopt.now.compose.ui.theme.NOWSOPTAndroidTheme
 
 class MainActivity : ComponentActivity() {
@@ -18,12 +16,5 @@ class MainActivity : ComponentActivity() {
                 MainScreen(usernameFromSignIn ?: "")
             }
         }
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-fun MainPreview() {
-    NOWSOPTAndroidTheme {
     }
 }
