+++ app/src/main/java/com/sopt/now/compose/feature/signIn/SignInActivity.kt
@@ -3,8 +3,6 @@ package com.sopt.now.compose.feature.signIn
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.tooling.preview.Preview
 import com.sopt.now.compose.ui.theme.NOWSOPTAndroidTheme
 
 class SignInActivity : ComponentActivity() {
@@ -15,13 +13,5 @@ class SignInActivity : ComponentActivity() {
                 SignInScreen()
             }
         }
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-private fun SignInPreview() {
-    NOWSOPTAndroidTheme {
-        SignInScreen()
     }
 }
