+++ app/src/main/java/com/sopt/now/compose/feature/signIn/SignInScreen.kt
@@ -22,6 +22,7 @@ import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.input.ImeAction
 import androidx.compose.ui.text.input.PasswordVisualTransformation
+import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.datastore.preferences.core.stringPreferencesKey
 import com.sopt.now.compose.R
@@ -29,6 +30,7 @@ import com.sopt.now.compose.feature.main.MainActivity
 import com.sopt.now.compose.feature.signUp.SignUpActivity
 import com.sopt.now.compose.feature.signUp.dataStore
 import com.sopt.now.compose.model.User
+import com.sopt.now.compose.ui.theme.NOWSOPTAndroidTheme
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.launch
 
@@ -45,20 +47,22 @@ fun SignInScreen() {
             value = username,
             onValueChange = { username = it },
             label = { Text(stringResource(id = R.string.username_label)) },
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(top = 16.dp),
-            keyboardOptions = KeyboardOptions(imeAction = ImeAction.Next)
+            modifier =
+                Modifier
+                    .fillMaxWidth()
+                    .padding(top = 16.dp),
+            keyboardOptions = KeyboardOptions(imeAction = ImeAction.Next),
         )
         OutlinedTextField(
             value = password,
             onValueChange = { password = it },
             label = { Text(stringResource(id = R.string.password_label)) },
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(top = 16.dp),
+            modifier =
+                Modifier
+                    .fillMaxWidth()
+                    .padding(top = 16.dp),
             visualTransformation = PasswordVisualTransformation(),
-            keyboardOptions = KeyboardOptions(imeAction = ImeAction.Done)
+            keyboardOptions = KeyboardOptions(imeAction = ImeAction.Done),
         )
         Button(
             onClick = {
@@ -67,7 +71,7 @@ fun SignInScreen() {
                         Toast.makeText(
                             context,
                             context.getString(R.string.sign_in_blank_error),
-                            Toast.LENGTH_SHORT
+                            Toast.LENGTH_SHORT,
                         ).show()
                         return@launch
                     }
@@ -78,11 +82,12 @@ fun SignInScreen() {
                         Toast.makeText(
                             context,
                             context.getText(R.string.sign_in_success),
-                            Toast.LENGTH_SHORT
+                            Toast.LENGTH_SHORT,
                         ).show()
-                        val intent = Intent(context, MainActivity::class.java).apply {
-                            putExtra("username", username)
-                        }
+                        val intent =
+                            Intent(context, MainActivity::class.java).apply {
+                                putExtra("username", username)
+                            }
                         context.startActivity(intent)
                         val activity = (context as SignInActivity)
                         activity.finish()
@@ -90,14 +95,15 @@ fun SignInScreen() {
                         Toast.makeText(
                             context,
                             context.getString(R.string.user_info_not_match),
-                            Toast.LENGTH_SHORT
+                            Toast.LENGTH_SHORT,
                         ).show()
                     }
                 }
             },
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(top = 16.dp)
+            modifier =
+                Modifier
+                    .fillMaxWidth()
+                    .padding(top = 16.dp),
         ) {
             Text("로그인")
         }
@@ -106,12 +112,21 @@ fun SignInScreen() {
                 val intent = Intent(context, SignUpActivity::class.java)
                 context.startActivity(intent)
             },
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(top = 8.dp)
+            modifier =
+                Modifier
+                    .fillMaxWidth()
+                    .padding(top = 8.dp),
         ) {
             Text(context.getString(R.string.sign_up_button))
         }
+    }
+}
+
+@Preview(showBackground = true)
+@Composable
+private fun SignInPreview() {
+    NOWSOPTAndroidTheme {
+        SignInScreen()
     }
 }
