+++ app/src/main/java/com/sopt/now/compose/feature/signUp/SignUpActivity.kt
@@ -3,8 +3,6 @@ package com.sopt.now.compose.feature.signUp
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.tooling.preview.Preview
 import com.sopt.now.compose.ui.theme.NOWSOPTAndroidTheme
 
 class SignUpActivity : ComponentActivity() {
@@ -15,13 +13,5 @@ class SignUpActivity : ComponentActivity() {
                 SignUpScreen()
             }
         }
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-private fun SignUpPreview() {
-    NOWSOPTAndroidTheme {
-        SignUpScreen()
     }
 }
