+++ app/src/main/java/com/sopt/now/compose/feature/signUp/SignUpScreen.kt
@@ -21,12 +21,14 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.input.ImeAction
 import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.text.input.PasswordVisualTransformation
+import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.datastore.preferences.core.edit
 import androidx.datastore.preferences.core.stringPreferencesKey
 import androidx.datastore.preferences.preferencesDataStore
 import com.sopt.now.compose.R
 import com.sopt.now.compose.model.User
+import com.sopt.now.compose.ui.theme.NOWSOPTAndroidTheme
 import kotlinx.coroutines.launch
 
 val Context.dataStore by preferencesDataStore(name = "userPreferences")
@@ -47,68 +49,77 @@ fun SignUpScreen() {
             value = username,
             onValueChange = { username = it },
             label = { Text(stringResource(id = R.string.username_label)) },
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(top = 16.dp),
-            keyboardOptions = KeyboardOptions(imeAction = ImeAction.Next)
+            modifier =
+                Modifier
+                    .fillMaxWidth()
+                    .padding(top = 16.dp),
+            keyboardOptions = KeyboardOptions(imeAction = ImeAction.Next),
         )
         OutlinedTextField(
             value = password,
             onValueChange = { password = it },
             label = { Text(stringResource(id = R.string.password_label)) },
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(top = 16.dp),
+            modifier =
+                Modifier
+                    .fillMaxWidth()
+                    .padding(top = 16.dp),
             visualTransformation = PasswordVisualTransformation(),
-            keyboardOptions = KeyboardOptions(imeAction = ImeAction.Next)
+            keyboardOptions = KeyboardOptions(imeAction = ImeAction.Next),
         )
         OutlinedTextField(
             value = nickname,
             onValueChange = { nickname = it },
             label = { Text(stringResource(id = R.string.nickname_label)) },
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(top = 16.dp),
-            keyboardOptions = KeyboardOptions(imeAction = ImeAction.Next)
+            modifier =
+                Modifier
+                    .fillMaxWidth()
+                    .padding(top = 16.dp),
+            keyboardOptions = KeyboardOptions(imeAction = ImeAction.Next),
         )
         OutlinedTextField(
             value = drinkCapacity,
             onValueChange = { drinkCapacity = it },
             label = { Text(stringResource(id = R.string.drink_capacity_label)) },
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(top = 16.dp),
-            keyboardOptions = KeyboardOptions(
-                keyboardType = KeyboardType.Number,
-                imeAction = ImeAction.Done
-            ),
+            modifier =
+                Modifier
+                    .fillMaxWidth()
+                    .padding(top = 16.dp),
+            keyboardOptions =
+                KeyboardOptions(
+                    keyboardType = KeyboardType.Number,
+                    imeAction = ImeAction.Done,
+                ),
         )
         Button(
             onClick = {
                 when {
-                    username.length !in MIN_USERNAME_LENGTH..MAX_USERNAME_LENGTH -> Toast.makeText(
-                        context,
-                        context.getString(R.string.username_error),
-                        Toast.LENGTH_SHORT
-                    ).show()
+                    username.length !in MIN_USERNAME_LENGTH..MAX_USERNAME_LENGTH ->
+                        Toast.makeText(
+                            context,
+                            context.getString(R.string.username_error),
+                            Toast.LENGTH_SHORT,
+                        ).show()
 
-                    password.length !in MIN_PASSWORD_LENGTH..MAX_PASSWORD_LENGTH -> Toast.makeText(
-                        context,
-                        context.getString(R.string.password_error),
-                        Toast.LENGTH_SHORT
-                    ).show()
+                    password.length !in MIN_PASSWORD_LENGTH..MAX_PASSWORD_LENGTH ->
+                        Toast.makeText(
+                            context,
+                            context.getString(R.string.password_error),
+                            Toast.LENGTH_SHORT,
+                        ).show()
 
-                    nickname.isBlank() -> Toast.makeText(
-                        context,
-                        context.getString(R.string.nickname_error),
-                        Toast.LENGTH_SHORT
-                    ).show()
+                    nickname.isBlank() ->
+                        Toast.makeText(
+                            context,
+                            context.getString(R.string.nickname_error),
+                            Toast.LENGTH_SHORT,
+                        ).show()
 
-                    drinkCapacity.isBlank() -> Toast.makeText(
-                        context,
-                        context.getString(R.string.drink_capacity_error),
-                        Toast.LENGTH_SHORT
-                    ).show()
+                    drinkCapacity.isBlank() ->
+                        Toast.makeText(
+                            context,
+                            context.getString(R.string.drink_capacity_error),
+                            Toast.LENGTH_SHORT,
+                        ).show()
 
                     else -> {
                         val user = User(username, password, nickname, drinkCapacity.toFloat())
@@ -116,19 +127,28 @@ fun SignUpScreen() {
                         Toast.makeText(
                             context,
                             context.getString(R.string.sign_up_success),
-                            Toast.LENGTH_SHORT
+                            Toast.LENGTH_SHORT,
                         ).show()
                         val activity = (context as SignUpActivity)
                         activity.finish()
                     }
                 }
             },
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(top = 16.dp)
+            modifier =
+                Modifier
+                    .fillMaxWidth()
+                    .padding(top = 16.dp),
         ) {
             Text(stringResource(id = R.string.sign_up_button))
         }
+    }
+}
+
+@Preview(showBackground = true)
+@Composable
+private fun SignUpPreview() {
+    NOWSOPTAndroidTheme {
+        SignUpScreen()
     }
 }
